requirements.txt
Flask
tweepy
python-dotenv